Improve commenting
